modules/uda-location-consumer/requirements.txt
kafka-python==2.0.2
psycopg2==2.9.1
protobuf==3.17.3
SQLAlchemy==1.3.19